+++ src/App.vue
@@ -1,38 +1,18 @@
 <template>
   <div id="app">
-    <button v-on:click="login">login</button>
+    <!-- <button v-on:click="login">login</button> -->
     <!-- <img alt="Vue logo" src="./assets/logo.png">
     <HelloWorld msg="Welcome to Your Vue.js App"/> -->
+    <CustomHeader/>
   </div>
 </template>
 
 <script>
-import firebase from 'firebase/app';
-import 'firebase/auth'
+import CustomHeader from './components/CustomHeader.vue'
 export default {
   name: 'App',
-  methods:{
-    login: function(){
-      console.log("login")
-      var provider = new firebase.auth.GithubAuthProvider();
-      firebase.auth().signInWithPopup(provider).then(function(result) {
-        // This gives you a GitHub Access Token. You can use it to access the GitHub API.
-        // var token = result.credential.accessToken;
-        // The signed-in user info.
-        var user = result.user;
-        console.log(user)
-        // ...
-      }).catch(function() {
-        // Handle Errors here.
-        // var errorCode = error.code;
-        // var errorMessage = error.message;
-        // The email of the user's account used.
-        // var email = error.email;
-        // The firebase.auth.AuthCredential type that was used.
-        // var credential = error.credential;
-        // ...
-      });
-    }
+  components: {
+    CustomHeader
   }
 }
 </script>
